feat: change php interpreter settings and update .gitignore
Remove .idea/php.xml as it contains user-specific settings, which should not be a part of version control. It might lead to problems when different contributors are working on the project with different PHP settings. Also, update the .gitignore file to ignore this file in the future.

Update the path of the executable file for Pest in .idea/php-test-framework.xml to the correct one, it will prevent execution problems in future tests.